requirements.txt
influxdb==2.10.0
PyYAML==3.11
